+++ fabric/src/main/java/io/github/mortuusars/exposure/fabric/ExposureFabric.java
@@ -24,14 +24,9 @@ public class ExposureFabric implements ModInitializer {
         Exposure.init();
 
         ModConfigEvents.reloading(Exposure.ID).register(config -> {
-            Config.reloading(config.getType());
             if (config.getType() == ModConfig.Type.COMMON && FabricLoader.getInstance().isModLoaded("create")) {
                 CreateFilmDeveloping.clearCachedData();
             }
-        });
-
-        ModConfigEvents.loading(Exposure.ID).register(config -> {
-            Config.loading(config.getType());
         });
 
         ModLoadingContext.registerConfig(Exposure.ID, ModConfig.Type.COMMON, Config.Common.SPEC);